Added the ability to input staff schedule occasions that recur.
Basically this is similar to the recurring bookings for resources
(meeting rooms and whatnot), but for staff schedules.  Adding this
feature motivated some minor refactoring, and a couple of other bug
fixes and small enhancements were picked up in passing.

Minute fields now list 00, 15, 30, and 45 first, before all the other
choices.  (Making this optional is deferred for now, because I want to
get the recurring staff schedule occasions checked in.)

When specifying recurring dates for bookings "as listed" (i.e.,
inputting individual dates one-by-one), the month gets pre-filled with
the current month initially; this can of course be changed.  (Making
it pre-fill to the same month as the previous listed date that was
just put in is deferred, again because I want to get recurring staff
schedule occasions checked in.)

A bug was fixed in the create-tables script that in some situations
might have caused extra colors to be added to the list of staff
schedule colors.