Per-site and per-resource customization of the auxilliary equipment offered when booking rooms.
This makes configurable (in the admin interface) one of the largest
remaining egregiously hardcoded things.  This will be beneficial even
for us here in Galion, as we will no longer be stuck offering the same
list of furniture and kitchen and A/V equipment for every room.  (The
list was originally constructed for our largest and most flexible
meeting room.  Some of it does not really apply to the other two
rooms.)  For other sites, this will be an even bigger deal, if you use
the room booking feature at all.