Allow local customization of the wording for certain things.
This is intended for things like buttons and links.  This first commit
just implements one link, the Personal Schedule link; but the
infrastructure for how it will work is determined now, so it won't be
hard to do more. All config variables for this will start with
wording_ and will be listed below all the other variables, so you can
easily ignore them unless there is some wording you want to customize.