ReSched Stats: correctly exclude auto-marked " Closed " bookings, by default.
If a resource was booked for "Closed", that booking was (by default)
excluded from the usage statistics; but the automatic closed-date
booking creates bookings for " Closed " instead (for reasons having
to do with being able to distinguish auto-bookings from manual ones),
and the stats gathering was not excluding these.  Now it does.