Program signup:  allow people to be signed up directly to the waiting list.
Optionally (depending on a new config variable), programs that are
allowed to have a waiting list can now allow people to be signed up
directly to the waiting list, even if the program is not full.

This change was requested by the Children's Room in Galion, so
that children who aren't technically in the right age range but
are close can be signed up to the waiting list just in case the
program doesn't fill up.