Staff Schedule: sanity checks to prevent bad data entry from causing server-side errors.
Previously, it was possible for a user to enter impossible times of
day (e.g., 28:00) in their staff schedule, which then caused errors
whenever anyone tried to view the staff schedule.

This fix does a couple of different things.  First, the display side
of things is more rubust now, so that even if invalid hours are
entered directly into the database, the whole schedule doesn't become
unviewable.

Second, there's also now a sanity check when the data are saved.

These fixes are for regular hours; I still need to look at event hours.