Email address validity check, and hide early start / late finish times when not in use.
When an email address is specified for sending meeting room policy, it is checked for
superficial validity immediately, so that e.g. if the user leaves off the top-level
domain suffix, they get immediate feedback.

To reduce confusion, the early start and late finish details (times and such)
are now only shown when the corresponding checkboxes are checked.