Initial implementation of the public self-signup feature for program signup.
This adds the public self-signup feature.  So far it is only lightly tested,
but we will be deploying it in Galion shortly, so it will get more testing
in the coming days and weeks.

Also, cleared up a couple of bugs that only existed this long because of a
missing "use strict" pragma, which has now been added.