Better error messages when DateTime cannot handle incorrect user input.
When using the DateTime module to construct a date/time object from user input,
do it inside eval, and give the user a human-readable error message explaining
the problem if it fails.

DateTime, for its part, handles bad input by just issuing a fatal error,
so traditionally the user would get Apache's generic server-side error message.
For an end user, this absolutely does not imply "Check your input, because
maybe there's no such thing as fifty-seven o'clock or February thirty-first".
I've been meaning to do something about this for a while, and finally got
around to it.

Also, while we're at it, explicity tell browsers that the charset is UTF-8.
Despite the fact that it's not 1998 any more, some browsers (*cough* Microsoft
*cough*) still assume other charsets by default.