Deep Freeze reboot tracking, and some cacheing work on usage stats.
Resources can now be marked with the D=DeepFreeze flag,
which cases rebooting them after each use to be tracked.
This works in basically the same way as the COVID-inspired
cleaned-after-each-use tracking.

Also, my working directory had a bunch of uncommitted
work on caching usage stats so they can be viewed with
something approaching reasonable performance.  This
work is not entirely complete, but it's better to have
it committed than not at this point.