Title view schedule page as first category that matches the resource list.
Previously, due to a lack-of-sort issue, if more than one category had
the same resource list, the page title on the view schedule page for
that category, was based on the name of the last-listed category that
matched, by default and coincidence.  Now it's based on the first-listed
one, and the code is deliberate and explicit about this.