Update MariaDB and Apache logos, Linux link.
The colors of the MariaDB logo have changed, presumably to make it
look less like the old MySQL dolphin, probably for reasons related
to trademark law.  In any case, we should be using the current logo.

While I'm at it, the Apache folks also updated their logo in 2019,
in this case I think more for aesthetic reasons, but we still may
as well be using the current logo.

Finally, linux.org no longer passes my sniff test, so link Tux
to the Wikipedia article on Linux instead.

This is all down in the footer where most users won't notice,
but still.