Don't offer timeslots before opening time in the morning.
This situation only arises if opening time is at different times
on different days of the week.  Because doview() is designed to
possibly show schedules for multiple days side by side in columns
of the same table, it creates rows starting from the earliest
start time specified by any of the schedules in play.

In the even that, in a given column, some of those rows occur
before the day's opening time, place "not open" text in the
resulting cell, rather than an "available" link and booking
button.