Generate correct redirect headers when creating a new program.
When a user creates a new program, in addition to creating the
program, ReSched also redirects to a signup sheet for that program.
This is all well and good, but the variables in the redirect header
were incorrectly handled in two different ways, resulting in a
signup sheet that did not actually work.  This is now fixed.